+++ git-updater-gist.php
@@ -13,7 +13,7 @@
  * Plugin Name:       GitHub Updater - Gist
  * Plugin URI:        https://github.com/afragen/git-updater-gist
  * Description:       Add GitHub Gist hosted repositories to the GitHub Updater plugin.
- * Version:           0.6.0
+ * Version:           0.6.0.1
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true
@@ -22,7 +22,7 @@
  * GitHub Plugin URI: https://github.com/afragen/git-updater-gist
  * Primary Branch:    main
  * Requires at least: 5.2
- * Requires PHP:      5.6
+ * Requires PHP:      7.0
  */
 
 namespace Fragen\Git_Updater\Gist;